change instance id data from 8 bit to 32 bit
(cherry picked from commit e51d584914f95f1e20840839bd360322299aa0fe)

+++ Source/ACE.Entity/Models/PropertiesPosition.cs
@@ -12,7 +12,7 @@ namespace ACE.Entity.Models
         public float RotationX { get; set; }
         public float RotationY { get; set; }
         public float RotationZ { get; set; }
-        public byte? Instance { get; set; }
+        public uint? Instance { get; set; }
 
         public PropertiesPosition Clone()
         {
